requirements.txt
streamlit
requests
pandas
python-dotenv
plotly
ollama
